feat(utils): move element to tree into utils
