+++ packages/utils/src/index.ts
@@ -1,5 +1,7 @@
 export * from "./tree-2-string";
 export * from "./types";
+export * from "./element-to-tree";
+export * from "./from-svg";
 export { styleToString } from "./style-to-string";
 export { mergeStyles } from "./merge-styles";
 export { handleClassName } from "./handle-class-name";
